+++ app/layout.tsx
@@ -4,6 +4,11 @@ import '../styles/globals.scss';
 import Footer from '../components/common/footer';
 import Header from '../components/common/header';
 
+// fontawesome
+import { config } from '@fortawesome/fontawesome-svg-core';
+import '@fortawesome/fontawesome-svg-core/styles.css';
+config.autoAddCss = false;
+
 // redux
 import { Provider } from 'react-redux';
 import { store } from '../store/store';
