+++ app/page.tsx
@@ -1,7 +1,9 @@
 import { v4 } from 'uuid';
 import HomePage from './HomePage';
 
-export default function Page() {
+const Page = () => {
   const roomId: string = v4();
   return <HomePage roomId={roomId} />;
-}
+};
+
+export default Page;
