+++ app/rooms/[roomId]/components/card.module.scss
