+++ app/rooms/[roomId]/components/table/summaryTags.tsx
@@ -1,7 +1,7 @@
 import { NextPage } from 'next';
-import { Card } from '../../../interfaces/card';
-import { Member } from '../../../interfaces/member';
-import { useAppSelector } from '../../../store/hooks';
+import { Card } from '../../../../../interfaces/card';
+import { Member } from '../../../../../interfaces/member';
+import { useAppSelector } from '../../../../../store/hooks';
 import SummaryTag from './summaryTag';
 
 const SummaryTags: NextPage = () => {
