+++ app/rooms/[roomId]/components/table/table.tsx
