+++ app/rooms/[roomId]/components/table/tableCard.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/color.scss';
+@import 'styles/color.scss';
 
 .blank {
   border: dashed 2px black;
