+++ app/rooms/[roomId]/components/tefuda/tefudaCard.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/color.scss';
+@import 'styles/color.scss';
 
 .selected {
   color: white;
