+++ package.json
@@ -20,9 +20,10 @@
     "@fortawesome/free-solid-svg-icons": "^6.2.0",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@reduxjs/toolkit": "^1.8.6",
+    "bufferutil": "^4.0.7",
     "bulma": "^0.9.4",
     "bulma-toast": "^2.4.1",
-    "express": "^4.18.1",
+    "express": "^4.18.2",
     "next": "^13.1.6",
     "next-redux-wrapper": "^8.0.0",
     "node-sass": "^7.0.3",
@@ -31,7 +32,8 @@
     "react-redux": "^8.0.4",
     "sass": "^1.55.0",
     "socket.io": "^4.5.2",
-    "socket.io-client": "^4.5.2",
+    "socket.io-client": "^4.5.4",
+    "utf-8-validate": "^6.0.2",
     "uuid": "^9.0.0"
   },
   "devDependencies": {
