+++ tsconfig.json
@@ -18,7 +18,8 @@
       {
         "name": "next"
       }
-    ]
+    ],
+    "baseUrl": "/"
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
   "exclude": ["node_modules"]
